Don't prevalidate keys
Instead, wait for a reply for the server indicating the arguments
were invalid. There's no point in validating the key in the client
when the server is doing so too. There's, however, a small tradeoff:
since we're using the binary protocol some keys which are not valid
when using the text protocol will be accepted.